Reply to a message by double clicking on its row
When a user double clicks on a message, or to the whitespace to its
side, reply to that message.

This shortcut makes is very fast and handy to quickly reply to a single
message and mimics what other popular messaging clients to (WhatsApp Web
and Telegram Desktop) are examples of this.